Moved get_class in models modules.
