+++ form_designer/views.py
@@ -8,28 +8,8 @@ from django.http import HttpResponseRedirect
 from django.conf import settings
 from django.contrib import messages
 from django.core.context_processors import csrf
-from django.core.exceptions import ImproperlyConfigured
-from django.utils.importlib import import_module
 
-from form_designer.models import FormDefinition
-
-def get_class(import_path):
-    try:
-        dot = import_path.rindex('.')
-    except ValueError:
-        raise ImproperlyConfigured("%s isn't a Python path." % import_path)
-    module, classname = import_path[:dot], import_path[dot + 1:]
-    try:
-        mod = import_module(module)
-    except ImportError, e:
-        raise ImproperlyConfigured('Error importing module %s: "%s"' %
-                                   (module, e))
-    try:
-        return getattr(mod, classname)
-    except AttributeError:
-        raise ImproperlyConfigured('Module "%s" does not define a "%s" '
-                                   'class.' % (module, classname))
-
+from form_designer.models import FormDefinition, get_class
 
 class DesignedForm(forms.Form):
     def __init__(self, form_definition, initial_data=None, *args, **kwargs):
